+++ docs/CHANGELOG.md
@@ -24,6 +24,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Fixed deployment error about missing warnings.
   ([#523](https://github.com/TheRacetrack/racetrack/issues/523))
+- Added informative error about crash during Job initialization.
+  ([#469](https://github.com/TheRacetrack/racetrack/issues/469))
 - Resolved compatibility issues with the Racetrack client
   for Python 3.8 related to unsupported typing operands.
   ([#527](https://github.com/TheRacetrack/racetrack/issues/527))